+++ src/django_elasticsearch_dsl_drf/filter_backends/filtering/__init__.py
@@ -1,0 +1,11 @@
+"""
+Term level filtering backends.
+"""
+
+from .common import *
+from .ids import *
+
+__title__ = 'django_elasticsearch_dsl_drf.filter_backends.filtering'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2016-2017 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'